refactor: update trading competition API integration for season 1 and season 2
